corrigiendo errores

+++ MicroservicioA/src/main/java/com/sv/microservicioa/service/VentasService.java
@@ -1,18 +1,14 @@
 package com.sv.microservicioa.service;
-
-import java.text.SimpleDateFormat;
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.env.Environment;
 import org.springframework.stereotype.Service;
-
 import com.sv.microservicioa.commons.AppConstans;
 import com.sv.microservicioa.dto.VentasDto;
 import com.sv.microservicioa.modelo.Empresas;
+import com.sv.microservicioa.modelo.Facturas;
+import com.sv.microservicioa.modelo.Totales;
 import com.sv.microservicioa.modelo.Ventas;
-import com.sv.microservicioa.repository.VentasRepository;
+import com.sv.microservicioa.repository.MicroserverArepository;
 import com.sv.microservicioa.util.DatosNoEncontradosException;
 
 @Service
@@ -22,106 +18,96 @@ public class VentasService {
 	@Autowired
 	private Environment env;
 	@Autowired
-	private VentasRepository vR;
+	private MicroserverArepository vR;
 
 	// 1. metodo para guardar:
 	public void guardarVentas(VentasDto vdto) {
 			try {
 				// 1. Respectivas validaciones:
-				if (vdto.getFecha() == null) {
-					throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
-							env.getProperty(AppConstans.ERROR_FECHA_MSG));
-						
-				} if (vdto.getEmpresasId() <= 0){
-					throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
-							env.getProperty(AppConstans.ERROR_IDEMPRESA_MSG));
-						
-				} if (vdto.getFactura() <= 0) { 
-					throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
-							env.getProperty(AppConstans.ERROR_FACTURA_MSG));
-						
-				} if (vdto.getFechaVenta() == null) {
-					throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
-							env.getProperty(AppConstans.ERROR_FECHAVENTA_MSG));
-					
-				} if (vdto.getFechaPago() == null){
-					throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
-							env.getProperty(AppConstans.ERROR_FECHAPAGO_MSG));
-						
-				} if (vdto.getImpaga() == null ||vdto.getImpaga().isEmpty()){ 
-					throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
-							env.getProperty(AppConstans.ERROR_IMPAGA_MSG));
-						
-				} if (vdto.getImpresa() == null ||vdto.getImpresa().isEmpty()) {
-					throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
-							env.getProperty(AppConstans.ERROR_IMPRESA_MSG));
-						
-				} if (vdto.getFormaPago() == null ||vdto.getFormaPago().isEmpty()){ 
-					throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
-							env.getProperty(AppConstans.ERROR_FORMAPAGO_MSG));
-					
-				} if (vdto.getTipoFactura() == null ||vdto.getTipoFactura().isEmpty()){ 
-					throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
-							env.getProperty(AppConstans.ERROR_TIPOFACTURA_MSG));
-						
-				} if(vdto.getSubTotal() == null || vdto.getSubTotal() <= 0){ 
-					throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
-							env.getProperty(AppConstans.ERROR_SUBTOTAL_MSG));
-						
-				} if(vdto.getSubTotalIva() == null || vdto.getSubTotalIva() <= 0) {
-					throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
-							env.getProperty(AppConstans.ERROR_IVA_MSG));
-						
-				} if(vdto.getSubTotalCesc() == null || vdto.getSubTotalCesc() <= 0) { 
-					throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
-							env.getProperty(AppConstans.ERROR_CESC_MSG));
-					
-				} if(vdto.getSubTotalDescuentos() == null || vdto.getSubTotalDescuentos() <= 0) {
-					throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
-							env.getProperty(AppConstans.ERROR_DESCUENTOS_MSG));
-						
-				} if (vdto.getTotalaPagar() == null ||vdto.getTotalaPagar() <= 0) { 
-					throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
-							env.getProperty(AppConstans.ERROR_TOTAL_MSG));
-				}
+//				if (vdto.getFecha() == null) {
+//					throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
+//							env.getProperty(AppConstans.ERROR_FECHA_MSG));
+//						
+//				} if (vdto.getEmpresasId() <= 0){
+//					throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
+//							env.getProperty(AppConstans.ERROR_IDEMPRESA_MSG));
+//						
+//				} if (vdto.getFactura() <= 0) { 
+//					throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
+//							env.getProperty(AppConstans.ERROR_FACTURA_MSG));
+//						
+//				} if (vdto.getFechaVenta() == null) {
+//					throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
+//							env.getProperty(AppConstans.ERROR_FECHAVENTA_MSG));
+//					
+//				} if (vdto.getFechaPago() == null){
+//					throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
+//							env.getProperty(AppConstans.ERROR_FECHAPAGO_MSG));
+//						
+//				} if (vdto.getImpaga() == null ||vdto.getImpaga().isEmpty()){ 
+//					throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
+//							env.getProperty(AppConstans.ERROR_IMPAGA_MSG));
+//						
+//				} if (vdto.getImpresa() == null ||vdto.getImpresa().isEmpty()) {
+//					throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
+//							env.getProperty(AppConstans.ERROR_IMPRESA_MSG));
+//						
+//				} if (vdto.getPagosId() <= 0){ 
+//					throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
+//							env.getProperty(AppConstans.ERROR_FORMAPAGO_MSG));
+//					
+//				} if (vdto.getTipoFactura() == null ||vdto.getTipoFactura().isEmpty()){ 
+//					throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
+//							env.getProperty(AppConstans.ERROR_TIPOFACTURA_MSG));
+//						
+//				} if(vdto.getSubTotal() == null || vdto.getSubTotal() <= 0){ 
+//					throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
+//							env.getProperty(AppConstans.ERROR_SUBTOTAL_MSG));
+//						
+//				} if(vdto.getSubTotalIva() == null || vdto.getSubTotalIva() <= 0) {
+//					throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
+//							env.getProperty(AppConstans.ERROR_IVA_MSG));
+//						
+//				} if(vdto.getSubTotalCesc() == null || vdto.getSubTotalCesc() <= 0) { 
+//					throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
+//							env.getProperty(AppConstans.ERROR_CESC_MSG));
+//					
+//				} if(vdto.getSubTotalDescuentos() == null || vdto.getSubTotalDescuentos() <= 0) {
+//					throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
+//							env.getProperty(AppConstans.ERROR_DESCUENTOS_MSG));
+//						
+//				} if (vdto.getTotalaPagar() == null ||vdto.getTotalaPagar() <= 0) { 
+//					throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
+//							env.getProperty(AppConstans.ERROR_TOTAL_MSG));
+//				}
 					
 
-				// 2. Instanciando mi entity para setiarla con los datos de la Dto:
+				// 2. Instanciando mis entitys para setiarlas con los datos de la Dto:
 				Ventas v = new Ventas();
-				v.setId(vdto.getId());
-				v.setFecha(vdto.getFecha());
-
-				// mandando a llamar la clase para setiar objeto(LLAVE FORANEA)
 				Empresas e = new Empresas();
-				e.setId(vdto.getEmpresasId());
+				Facturas f = new Facturas();
+				Totales p = new Totales();
+				
+				v.setFecha(vdto.getFecha());
+				e.setId(vdto.getEmpresa());
 				v.setEmpresasId(e);
-
-				v.setFactura(vdto.getFactura());
-				/*
-				 * // convertir String a Date: FECHA DE LA VENTA SimpleDateFormat formato = new
-				 * SimpleDateFormat("yyyy-MM-dd"); Date fechaDate = null; fechaDate =
-				 * formato.parse(vdto.getFechaVenta()); v.setFechaVenta(fechaDate);
-				 */
+				f.setNumeroFactura(vdto.getFactura());
+				v.setFactura(f);
 				v.setFechaVenta(vdto.getFechaVenta());
-				/*
-				 * //Convertir String a Date : FECHA DE PAGO SimpleDateFormat formato1 = new
-				 * SimpleDateFormat("yyyy-MM-dd"); Date fechaDate2 = null; fechaDate2 =
-				 * formato1.parse(vdto.getFechaPago()); v.setFechaPago(fechaDate2);
-				 */
-				v.setFechaPago(vdto.getFechaPago());
-				v.setImpaga(vdto.getImpaga());
-				v.setImpresa(vdto.getImpresa());
-				v.setFormaPago(vdto.getFormaPago());
-				v.setTipoFactura(vdto.getTipoFactura());
-				v.setSubTotal(vdto.getSubTotal());
-				v.setSubTotalIva(vdto.getSubTotalIva());
-				v.setSubTotalCesc(vdto.getSubTotalCesc());
-				v.setSubTotalDescuentos(vdto.getSubTotalDescuentos());
-				v.setTotalaPagar(vdto.getTotalaPagar());
+				p.setFechapago(vdto.getFechaPago());
+				p.setImpaga(vdto.getImpaga());
+				f.setImpresa(vdto.getImpresa());
+				p.setFormaPago(vdto.getFechaPago());
+				f.setTipoFactura(vdto.getTipoFactura());
+				p.setSubTotal(vdto.getSubTotal());
+				p.setSubtoTaliva(vdto.getSubTotalIva());
+				p.setSubTotalcesc(vdto.getSubTotalCesc());
+				p.setSubTotaldescuentos(vdto.getSubTotalDescuentos());
+				p.setTotalaPagar(vdto.getTotalaPagar());
 
-				if (v.getFactura() != 0) {
+				if (v.getFactura() != null) {
 					// validacion si la nomina ya esta registrada:
-					Ventas vv = vR.buscarSiYaExiste(v.getFactura());
+					Facturas vv = vR.buscarSiYaExiste(v.getFactura());
 
 					if (vv != null) {
 						throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
@@ -144,86 +130,87 @@ public class VentasService {
 	public void EditarVentas(VentasDto vdto) {
 		try {
 			// 1. Respectivas validaciones:
-			if (vdto.getFecha() == null) {
-				throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
-						env.getProperty(AppConstans.ERROR_FECHA_MSG));
-					
-			} if (vdto.getEmpresasId() <= 0){
-				throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
-						env.getProperty(AppConstans.ERROR_IDEMPRESA_MSG));
-					
-			} if (vdto.getFactura() <= 0) { 
-				throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
-						env.getProperty(AppConstans.ERROR_FACTURA_MSG));
-					
-			} if (vdto.getFechaVenta() == null) {
-				throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
-						env.getProperty(AppConstans.ERROR_FECHAVENTA_MSG));
-				
-			} if (vdto.getFechaPago() == null){
-				throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
-						env.getProperty(AppConstans.ERROR_FECHAPAGO_MSG));
-					
-			} if (vdto.getImpaga() == null ||vdto.getImpaga().isEmpty()){ 
-				throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
-						env.getProperty(AppConstans.ERROR_IMPAGA_MSG));
-					
-			} if (vdto.getImpresa() == null ||vdto.getImpresa().isEmpty()) {
-				throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
-						env.getProperty(AppConstans.ERROR_IMPRESA_MSG));
-					
-			} if (vdto.getFormaPago() == null ||vdto.getFormaPago().isEmpty()){ 
-				throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
-						env.getProperty(AppConstans.ERROR_FORMAPAGO_MSG));
-				
-			} if (vdto.getTipoFactura() == null ||vdto.getTipoFactura().isEmpty()){ 
-				throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
-						env.getProperty(AppConstans.ERROR_TIPOFACTURA_MSG));
-					
-			} if(vdto.getSubTotal() == null || vdto.getSubTotal() <= 0){ 
-				throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
-						env.getProperty(AppConstans.ERROR_SUBTOTAL_MSG));
-					
-			} if(vdto.getSubTotalIva() == null || vdto.getSubTotalIva() <= 0) {
-				throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
-						env.getProperty(AppConstans.ERROR_IVA_MSG));
-					
-			} if(vdto.getSubTotalCesc() == null || vdto.getSubTotalCesc() <= 0) { 
-				throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
-						env.getProperty(AppConstans.ERROR_CESC_MSG));
-				
-			} if(vdto.getSubTotalDescuentos() == null || vdto.getSubTotalDescuentos() <= 0) {
-				throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
-						env.getProperty(AppConstans.ERROR_DESCUENTOS_MSG));
-					
-			} if (vdto.getTotalaPagar() == null ||vdto.getTotalaPagar() <= 0) { 
-				throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
-						env.getProperty(AppConstans.ERROR_TOTAL_MSG));
-			}
+//			if (vdto.getFecha() == null) {
+//				throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
+//						env.getProperty(AppConstans.ERROR_FECHA_MSG));
+//					
+//			} if (vdto.getEmpresasId() <= 0){
+//				throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
+//						env.getProperty(AppConstans.ERROR_IDEMPRESA_MSG));
+//					
+//			} if (vdto.getFactura() <= 0) { 
+//				throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
+//						env.getProperty(AppConstans.ERROR_FACTURA_MSG));
+//					
+//			} if (vdto.getFechaVenta() == null) {
+//				throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
+//						env.getProperty(AppConstans.ERROR_FECHAVENTA_MSG));
+//				
+//			} if (vdto.getFechaPago() == null){
+//				throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
+//						env.getProperty(AppConstans.ERROR_FECHAPAGO_MSG));
+//					
+//			} if (vdto.getImpaga() == null ||vdto.getImpaga().isEmpty()){ 
+//				throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
+//						env.getProperty(AppConstans.ERROR_IMPAGA_MSG));
+//					
+//			} if (vdto.getImpresa() == null ||vdto.getImpresa().isEmpty()) {
+//				throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
+//						env.getProperty(AppConstans.ERROR_IMPRESA_MSG));
+//					
+//			} if (vdto.getFormaPago() == null ||vdto.getFormaPago().isEmpty()){ 
+//				throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
+//						env.getProperty(AppConstans.ERROR_FORMAPAGO_MSG));
+//				
+//			} if (vdto.getTipoFactura() == null ||vdto.getTipoFactura().isEmpty()){ 
+//				throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
+//						env.getProperty(AppConstans.ERROR_TIPOFACTURA_MSG));
+//					
+//			} if(vdto.getSubTotal() == null || vdto.getSubTotal() <= 0){ 
+//				throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
+//						env.getProperty(AppConstans.ERROR_SUBTOTAL_MSG));
+//					
+//			} if(vdto.getSubTotalIva() == null || vdto.getSubTotalIva() <= 0) {
+//				throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
+//						env.getProperty(AppConstans.ERROR_IVA_MSG));
+//					
+//			} if(vdto.getSubTotalCesc() == null || vdto.getSubTotalCesc() <= 0) { 
+//				throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
+//						env.getProperty(AppConstans.ERROR_CESC_MSG));
+//				
+//			} if(vdto.getSubTotalDescuentos() == null || vdto.getSubTotalDescuentos() <= 0) {
+//				throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
+//						env.getProperty(AppConstans.ERROR_DESCUENTOS_MSG));
+//					
+//			} if (vdto.getTotalaPagar() == null ||vdto.getTotalaPagar() <= 0) { 
+//				throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
+//						env.getProperty(AppConstans.ERROR_TOTAL_MSG));
+//			}
 				
 
 			// 2. Instanciando mi entity para setiarla con los datos de la Dto:
 			Ventas v = new Ventas();
-			v.setId(vdto.getId());
-			v.setFecha(vdto.getFecha());
-
-			// mandando a llamar la clase para setiar objeto(LLAVE FORANEA)
 			Empresas e = new Empresas();
-			e.setId(vdto.getEmpresasId());
+			Facturas f = new Facturas();
+			Totales p = new Totales();
+			
+			v.setFecha(vdto.getFecha());
+			e.setId(vdto.getEmpresa());
 			v.setEmpresasId(e);
-
-			v.setFactura(vdto.getFactura());
+			f.setNumeroFactura(vdto.getFactura());
+			v.setFactura(f);
 			v.setFechaVenta(vdto.getFechaVenta());
-			v.setFechaPago(vdto.getFechaPago());
-			v.setImpaga(vdto.getImpaga());
-			v.setImpresa(vdto.getImpresa());
-			v.setFormaPago(vdto.getFormaPago());
-			v.setTipoFactura(vdto.getTipoFactura());
-			v.setSubTotal(vdto.getSubTotal());
-			v.setSubTotalIva(vdto.getSubTotalIva());
-			v.setSubTotalCesc(vdto.getSubTotalCesc());
-			v.setSubTotalDescuentos(vdto.getSubTotalDescuentos());
-			v.setTotalaPagar(vdto.getTotalaPagar());
+			p.setFechapago(vdto.getFechaPago());
+			p.setImpaga(vdto.getImpaga());
+			f.setImpresa(vdto.getImpresa());
+			p.setFormaPago(vdto.getFechaPago());
+			f.setTipoFactura(vdto.getTipoFactura());
+			p.setSubTotal(vdto.getSubTotal());
+			p.setSubtoTaliva(vdto.getSubTotalIva());
+			p.setSubTotalcesc(vdto.getSubTotalCesc());
+			p.setSubTotaldescuentos(vdto.getSubTotalDescuentos());
+			p.setTotalaPagar(vdto.getTotalaPagar());
+
 			vR.save(v);
 			
 		} catch (DatosNoEncontradosException exc) {
@@ -242,7 +229,15 @@ public class VentasService {
 			if(id <= 0 || id == null ) {
 			throw new DatosNoEncontradosException("ERROR EL ID NO EXISTE");
 		}
+			Ventas v = new Ventas();
+			Facturas vv = vR.buscarSiYaExiste(v.getFactura());
+
+		if (vv != null) {
+			throw new DatosNoEncontradosException(env.getProperty(AppConstans.ERROR_CAMPOVACIO_COD),
+					env.getProperty(AppConstans.ERROR_GUARDAR_MSG));
+		} else {
 			vR.deleteById(id);
+		}
 		}catch (DatosNoEncontradosException exc ) {
 			throw exc;
 		}catch (Exception e) {
@@ -289,14 +284,14 @@ public class VentasService {
 			 iva = 0.13;
 			 totalIva = totalizarSubT() * iva;
 			 totalCesc = totalizarSubT() * cesc;
-			 total = totalizarSubT() + totalIva * cesc;
-			 utilidades = total - (totalIva + cesc);
+			 total = totalizarSubT() + totalIva + totalCesc;
+			 utilidades = totalIva  + totalCesc;
 			 
 		}catch(Exception e) {
 			e.printStackTrace();
 		} 
 		return "Subtotal: $" + totalizarSubT() + " , Iva: $" + totalIva +
-				" , Cesc: $" + totalCesc + " , Total: $" + total + ", Utilidades: $" + utilidades;  
+				" , Cesc: $" + totalCesc + ", Utilidades: $" + utilidades + " , Total: $" + total;  
 		
 	}  // fin del metodo imprimir calculos de ley.
